Implemented basic functionality to let user specify testcases to run on the command line:
venus exec --test=foo

If 'foo.js' exists in the current directory, then it will be loaded.

If 'foo' is a directory under the current directory, then all of its
children (recursively) will be loaded.

When a test is loaded, it is assigned an id, and the config comments in
the test file (if there are any) are parsed.

Also added a very bare-bones version of the /test/:testid route to the
executor server.