Selenium RC integration now works. Some new commandline flags added:
`venus run --test spec.js --selenium --selenium-server 127.1.2.3 --selenium-browser chrome`

This will instruct venus to run the spec.js test on a selenium RC server (or grid) located at 127.1.2.3. It will also request Chrome as the browser to use.

Also, added feature for browser name and version to be printed with test results in the CLI output.