Mocha & IE bug fix
In IE, the mocha library was executing tests before we could register to listen for the test result events.
Thus, our adaptor was not getting notified of the test results. Had to hack the mocha library a bit. Will investigate more
and see if this is a general bug in mocha that we should report.