Improved performance of running with PhantomJS
When running with GhostDriverUac (to execute tests with phantomjs), we were previously spawning the phantomjs process and then waiting several seconds before proceeding.
This was intended to give the phantomjs process enough time to spin up on slower machines.

With this change, rather than waiting an arbitrary length of time, we poll every 50ms to see if the process is up by sending an http request to the server which the process spawns.
After 300 attempts, we give up.